DragonFly has different stat, dirent, clock_t, ino_t, nlink_t, blksize_t
Move FreeBSD's stat into freebsd_{x86,x86_64}.rs.
